# Conflicts:
#	src/main/java/ua/chernonog/users/controller/UserController.java

+++ src/main/java/ua/chernonog/users/service/UserService.java
@@ -1,13 +1,13 @@
 package ua.chernonog.users.service;
 
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
+import ua.chernonog.users.config.UsersProps;
 import ua.chernonog.users.entity.UserEntity;
+import ua.chernonog.users.exception.UnderAgeException;
+import ua.chernonog.users.exception.UserNotFoundException;
 import ua.chernonog.users.mapper.UserCustomMapper;
 import ua.chernonog.users.mapper.UserMapper;
 import ua.chernonog.users.model.request.BirthdateRangeRequest;
@@ -18,7 +18,6 @@ import ua.chernonog.users.repository.UserRepository;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Slf4j
 @AllArgsConstructor
@@ -26,53 +25,49 @@ import java.util.stream.Collectors;
 @Transactional
 
 public class UserService {
-    UserRepository userRepository;
-    UserMapper userMapper;
-    UserCustomMapper userCustomMapper;
+    private UserRepository userRepository;
+    private UserMapper userMapper;
+    private UserCustomMapper userCustomMapper;
+    private UsersProps usersProps;
 
-//
-public UserResponse saveUser(UserRequest userRequest) {
-    if (!IsNotLessThan18Age(userRequest)) {
-        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-    } else {
-        UserEntity userEntity = userRepository.save(UserEntity.builder()
-                .email(userRequest.getEmail())
-                .firstName(userRequest.getFirstName())
-                .lastName(userRequest.getLastName())
-                .address(userRequest.getAddress())
-                .phoneNumber(userRequest.getPhoneNumber())
-                .birthdate(userRequest.getBirthdate())
-                .build());
-        userMapper.userEntityToUserResponse(userEntity);
-        return userMapper.userEntityToUserResponse(userEntity);
-    }
+    public UserResponse saveUser(UserRequest userRequest) {
+        if (isLessThan18Age(userRequest)) {
+            throw new UnderAgeException();
+        } else {
+            UserEntity savedUserEntity = userRepository.save(UserEntity.builder()
+                    .email(userRequest.getEmail())
+                    .firstName(userRequest.getFirstName())
+                    .lastName(userRequest.getLastName())
+                    .address(userRequest.getAddress())
+                    .phoneNumber(userRequest.getPhoneNumber())
+                    .birthdate(userRequest.getBirthdate())
+                    .build());
+            userMapper.userEntityToUserResponse(savedUserEntity);
+            return userMapper.userEntityToUserResponse(savedUserEntity);
+        }
     }
 
-    private boolean IsNotLessThan18Age(UserRequest userRequest) {
+    private boolean isLessThan18Age(UserRequest userRequest) {
         LocalDate currentDate = LocalDate.now();
         LocalDate birthdate = userRequest.getBirthdate();
-
-        // Визначаємо різницю між поточною датою та датою народження
         Period age = Period.between(birthdate, currentDate);
-
-        return age.getYears() >= 18;
+        return age.getYears() < usersProps.legalAge();
     }
 
     public UserResponse updateUser(long id, UserRequest userRequest) {
-
         UserEntity foundedUser = userRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND));
-
+                UserNotFoundException::new);
         UserEntity updatedUser = userRepository.save(userCustomMapper.updateUserFields(foundedUser, userRequest));
         return userMapper.userEntityToUserResponse(updatedUser);
-
     }
 
     public String deleteUser(long id) {
-        userRepository.deleteById(id);
-
-
-        return "User deleted successful";
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+            return "User deleted successfully";
+        } else {
+            throw new UserNotFoundException();
+        }
     }
 
     @Transactional(readOnly = true)
@@ -80,11 +75,11 @@ public UserResponse saveUser(UserRequest userRequest) {
         List<UserEntity> foundedUsers = userRepository.findByBirthdateBetween(birthdateRangeRequest.getFrom(), birthdateRangeRequest.getTo());
         return foundedUsers.stream().map(userEntity ->
                 userMapper.userEntityToUserResponse(userEntity)).toList();
-
     }
 
+    @Transactional(readOnly = true)
     public List<UserResponse> findAllUsers() {
-      return userRepository.findAll().stream().map(userMapper::userEntityToUserResponse).collect(Collectors.toList());
+        return userRepository.findAll().stream().map(userMapper::userEntityToUserResponse).toList();
     }
     public String greet() {
         return "Hello, World";
